% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChadLM.R
\name{ChadLM}
\alias{ChadLM}
\title{ChadLM}
\usage{
ChadLM(y, X, lambda)
}
\arguments{
\item{y}{respeonse vector}

\item{X}{covariates}

\item{lambda}{penalization parameter}
}
\value{
overestimated effect sizes
}
\description{
ChadLM
}
